newcastle single read usable

+++ functions.py
@@ -52,12 +52,13 @@ def run_loop(arg_options):  #calls read_aligner
 
     list_of_files = glob.glob('*gz')
     list_len = len(list_of_files)
-    if (list_len % 2 != 0):
-        print("\n#####Check paired files.  Unpaired files seen by odd number of counted FASTQs\n\n")
-        sys.exit(0)
+    # if (list_len % 2 != 0):
+    #     print("\n#####Check paired files.  Unpaired files seen by odd number of counted FASTQs\n\n")
+    #     sys.exit(0)
 
     for afile in list_of_files:
         prefix_name = re.sub('_.*', '', afile)
+        prefix_name = re.sub('\..*', '', prefix_name)
         print(prefix_name)
         if not os.path.exists(prefix_name):
             os.makedirs(prefix_name)
@@ -236,10 +237,19 @@ def run_loop(arg_options):  #calls read_aligner
 
 def read_aligner(sample_name, arg_options):
 
+    os.chdir(arg_options['root_dir'] + "/" + sample_name)
     sample_directory = str(os.getcwd())
-    os.chdir(sample_name)
-    R1 = glob.glob('*_R1*fastq.gz')
-    R2 = glob.glob('*_R2*fastq.gz')
+
+    pair_check = len(glob.glob('*_R2*fastq.gz'))
+    if pair_check > 0:
+        R1 = glob.glob('*_R1*fastq.gz')
+        R2 = glob.glob('*_R2*fastq.gz')
+        paired = True
+    else:
+        R1 = glob.glob('*fastq.gz')
+        R2 = None
+        paired = False
+    arg_options['paired'] =  paired
 
     if len(R1) > 1:
         print("#### Check for a duplicate file in {}" .format(sample_name))
@@ -247,9 +257,10 @@ def read_aligner(sample_name, arg_options):
 
     os.makedirs("zips")
     shutil.move(R1[0], "zips")
-    shutil.move(R2[0], "zips")
-    arg_options['R1'] = sample_directory + "/" + sample_name + "/zips/" + R1[0]
-    arg_options['R2'] = sample_directory + "/" + sample_name + "/zips/" + R2[0]
+    arg_options['R1'] = sample_directory + "/zips/" + R1[0]
+    if paired:
+        shutil.move(R2[0], "zips")
+        arg_options['R2'] = sample_directory + "/zips/" + R2[0]
 
     read_quality_stats = {}
     print("Getting mean for {}" .format(arg_options['R1']))
@@ -263,17 +274,18 @@ def read_aligner(sample_name, arg_options):
     thirty_or_greater_count = sum(i > 29 for i in mean_quality_list)
     read_quality_stats["Q30_R1"] = "{:.1%}" .format(thirty_or_greater_count / len(mean_quality_list))
 
-    print("Getting mean for {}" .format(arg_options['R2']))
-    handle = gzip.open(arg_options['R2'], "rt")
-    mean_quality_list = []
-    for rec in SeqIO.parse(handle, "fastq"):
-        mean_q = get_read_mean(rec)
-        mean_quality_list.append(mean_q)
+    if paired:
+        print("Getting mean for {}" .format(arg_options['R2']))
+        handle = gzip.open(arg_options['R2'], "rt")
+        mean_quality_list = []
+        for rec in SeqIO.parse(handle, "fastq"):
+            mean_q = get_read_mean(rec)
+            mean_quality_list.append(mean_q)
 
-    read_quality_stats["Q_ave_R2"] = "{:.1f}" .format(mean(mean_quality_list))
-    thirty_or_greater_count = sum(i > 29 for i in mean_quality_list)
-    read_quality_stats["Q30_R2"] = "{:.1%}" .format(thirty_or_greater_count / len(mean_quality_list))
-    arg_options['read_quality_stats'] = read_quality_stats
+        read_quality_stats["Q_ave_R2"] = "{:.1f}" .format(mean(mean_quality_list))
+        thirty_or_greater_count = sum(i > 29 for i in mean_quality_list)
+        read_quality_stats["Q30_R2"] = "{:.1%}" .format(thirty_or_greater_count / len(mean_quality_list))
+        arg_options['read_quality_stats'] = read_quality_stats
 
     arg_options['sample_name'] = sample_name
     arg_options = species_selection_step1(arg_options)
@@ -489,22 +501,25 @@ def finding_best_ref(v, fastq_list):
 
 
 def align_reads(arg_options):
+    paired = arg_options['paired']
     working_directory = os.getcwd()
     print("Working on: {}" .format(working_directory))
     ts = time.time()
     st = datetime.fromtimestamp(ts).strftime('%Y-%m-%d_%H-%M-%S')
     R1 = arg_options['R1']
-    R2 = arg_options['R2']
+    if paired:
+        R2 = arg_options['R2']
     if arg_options["species"] is None:
-        R1size = sizeof_fmt(os.path.getsize(R1))
-        R2size = sizeof_fmt(os.path.getsize(R2))
         stat_summary = {}
+        R1size = sizeof_fmt(os.path.getsize(R1))
+        if paired:
+            R2size = sizeof_fmt(os.path.getsize(R2))
+            stat_summary["R2size"] = R2size
         stat_summary["time_stamp"] = st
         stat_summary["sample_name"] = arg_options["sample_name"]
         stat_summary["species"] = "NOT_FOUND"
         stat_summary["reference_sequence_name"] = "N/A"
         stat_summary["R1size"] = R1size
-        stat_summary["R2size"] = R2size
         stat_summary["allbam_mapped_reads"] = "CHECK SAMPLE *****************************************"
         stat_summary.update(arg_options['read_quality_stats'])
         return(stat_summary)
@@ -535,7 +550,8 @@ def align_reads(arg_options):
         print("sample name: %s" % sample_name)
         print("sample reference: %s" % sample_reference)
         print("Read 1: %s" % R1)
-        print("Read 2: %s\n" % R2)
+        if paired:
+            print("Read 2: %s\n" % R2)
         print("working_directory: %s" % working_directory)
         print("--")
 
@@ -549,9 +565,10 @@ def align_reads(arg_options):
         unmapsam = loc_sam + "-unmapped.sam"
         metrics = loc_sam + "-metrics.txt"
         unmapped_read1 = loc_sam + "-unmapped_R1.fastq"
-        unmapped_read2 = loc_sam + "-unmapped_R2.fastq"
         unmapped_read1gz = loc_sam + "-unmapped_R1.fastq.gz"
-        unmapped_read2gz = loc_sam + "-unmapped_R2.fastq.gz"
+        if paired:
+            unmapped_read2 = loc_sam + "-unmapped_R2.fastq"
+            unmapped_read2gz = loc_sam + "-unmapped_R2.fastq.gz"
         sortedbam = loc_sam + "-sorted.bam"
         nodupbam = loc_sam + "-nodup.bam"
         unfiltered_hapall = loc_sam + "-unfiltered_hapall.vcf"
@@ -560,8 +577,11 @@ def align_reads(arg_options):
         zero_coverage_vcf = loc_sam + "_zc.vcf"
 
         #########################################################
-        print("\n@@@ BWA mem: {}" .format(sample_name))
-        os.system(r'bwa mem -M -R "@RG\tID:%s\tSM:%s\tPL:ILLUMINA\tPI:250" -t 16 %s %s %s > %s' % (sample_name, sample_name, sample_reference, R1, R2, samfile))
+        print("\n@@@ BWA mem: {}".format(sample_name))
+        if paired:
+            os.system(r'bwa mem -M -R "@RG\tID:%s\tSM:%s\tPL:ILLUMINA\tPI:250" -t 16 %s %s %s > %s' % (sample_name, sample_name, sample_reference, R1, R2, samfile))
+        else:
+            os.system(r'bwa mem -M -R "@RG\tID:%s\tSM:%s\tPL:ILLUMINA\tPI:250" -t 16 %s %s > %s' % (sample_name, sample_name, sample_reference, R1, samfile))
         os.system("samtools view -Sb {} -o {}" .format(samfile, bamfile))
         os.system("samtools sort {} -o {}" .format(bamfile, sortedbam))
         os.system("samtools index {}" .format(sortedbam))
@@ -602,15 +622,24 @@ def align_reads(arg_options):
         os.system(r'vcffilter -f "QUAL > 20" %s > %s' % (mapq_fix, hapall))
 
         print("\n@@@ Assemble Unmapped Reads: {}" .format(sample_name))
-        os.system("samtools view -h -f4 -T {} {} -o {}" .format(sample_reference, nodupbam, unmapsam))
-        os.system("picard SamToFastq INPUT={} FASTQ={} SECOND_END_FASTQ={}" .format(unmapsam, unmapped_read1, unmapped_read2))
+        os.system("samtools view -h -f4 -T {} {} -o {}".format(sample_reference, nodupbam, unmapsam))
+        if paired:
+            os.system("picard SamToFastq INPUT={} FASTQ={} SECOND_END_FASTQ={}".format(unmapsam, unmapped_read1, unmapped_read2))
+        else:
+            os.system("picard SamToFastq INPUT={} FASTQ={}".format(unmapsam, unmapped_read1))
 
         abyss_contig_count = 0
         try:
-            os.system("ABYSS --out {} --coverage 5 --kmer 64 {} {}" .format(abyss_out, unmapped_read1, unmapped_read2))
-            with open(abyss_out) as f:
-                for line in f:
-                    abyss_contig_count += line.count(">")
+            if paired:
+                os.system("ABYSS --out {} --coverage 5 --kmer 64 {} {}" .format(abyss_out, unmapped_read1, unmapped_read2))
+                with open(abyss_out) as f:
+                    for line in f:
+                        abyss_contig_count += line.count(">")
+            else:
+                os.system("ABYSS --out {} --coverage 5 --kmer 64 {}" .format(abyss_out, unmapped_read1))
+                with open(abyss_out) as f:
+                    for line in f:
+                        abyss_contig_count += line.count(">")
         except FileNotFoundError:
             abyss_contig_count = 0
 
@@ -656,7 +685,7 @@ def align_reads(arg_options):
             smtp.send_message(msg)
             smtp.quit()
 
-        if arg_options["gbk_file"] is not "None" and not arg_options['no_annotation']:
+        if arg_options["gbk_file"] or arg_options['no_annotation']:
             annotated_vcf = loc_sam + "-annotated.vcf"
             gbk_file = arg_options['gbk_file']
 
@@ -735,10 +764,12 @@ def align_reads(arg_options):
                 for cf in cat_files:
                     with open(cf, "rb") as infile:
                         outfile.write(infile.read())
-
-        os.remove('temp.csv')
-        os.remove('v_header.csv')
-        os.remove('v_annotated_body.csv')
+        try:
+            os.remove('temp.csv')
+            os.remove('v_header.csv')
+            os.remove('v_annotated_body.csv')
+        except FileNotFoundError:
+            pass
         os.remove(samfile)
         os.remove(bamfile)
         os.remove(unmapsam)
@@ -759,14 +790,17 @@ def align_reads(arg_options):
         newZip = zipfile.ZipFile(unmapped_read1gz, 'w')
         newZip.write(unmapped_read1, compress_type=zipfile.ZIP_DEFLATED)
         newZip.close()
-        newZip = zipfile.ZipFile(unmapped_read2gz, 'w')
-        newZip.write(unmapped_read2, compress_type=zipfile.ZIP_DEFLATED)
-        newZip.close()
         os.remove(unmapped_read1)
-        os.remove(unmapped_read2)
+        if paired:
+            newZip = zipfile.ZipFile(unmapped_read2gz, 'w')
+            newZip.write(unmapped_read2, compress_type=zipfile.ZIP_DEFLATED)
+            newZip.close()
+            os.remove(unmapped_read2)
+        
         try:
             shutil.move(unmapped_read1gz, unmapped)
-            shutil.move(unmapped_read2gz, unmapped)
+            if paired:
+                shutil.move(unmapped_read2gz, unmapped)
             shutil.move(abyss_out, unmapped)
         except FileNotFoundError:
             pass
@@ -791,7 +825,8 @@ def align_reads(arg_options):
         print("average_coverage: %s" % ave_coverage)
 
         R1size = sizeof_fmt(os.path.getsize(R1))
-        R2size = sizeof_fmt(os.path.getsize(R2))
+        if paired:
+            R2size = sizeof_fmt(os.path.getsize(R2))
 
         try:
             with open("mlst/mlst.txt") as f:
@@ -832,7 +867,7 @@ def align_reads(arg_options):
 
         sequence_count = 0
         total_length = 0
-        with gzip.open(R2, "rt") as handle:
+        with gzip.open(R1, "rt") as handle:
             for r in SeqIO.parse(handle, "fastq"):
                 total_length = total_length + len(r.seq)
                 sequence_count = sequence_count + 1
@@ -848,7 +883,10 @@ def align_reads(arg_options):
         stat_summary["species"] = arg_options["species"]
         stat_summary["reference_sequence_name"] = reference_sequence_name
         stat_summary["R1size"] = R1size
-        stat_summary["R2size"] = R2size
+        if paired:
+            stat_summary["R2size"] = R2size
+        else:
+            stat_summary["R2size"] = None
         stat_summary["allbam_mapped_reads"] = allbam_mapped_reads
         stat_summary["genome_coverage"] = genome_coverage
         stat_summary["ave_coverage"] = ave_coverage
@@ -874,7 +912,10 @@ def align_reads(arg_options):
             worksheet.write(row, col, header)
             col += 1
             # worksheet.write(row, col, v)
-        stat_summary.update(arg_options['read_quality_stats'])
+        try:
+            stat_summary.update(arg_options['read_quality_stats'])
+        except KeyError:
+            pass
         worksheet.write(1, 0, stat_summary.get('time_stamp', 'n/a'))
         worksheet.write(1, 1, stat_summary.get('sample_name', 'n/a'))
         worksheet.write(1, 2, stat_summary.get('species', 'n/a'))
@@ -1337,40 +1378,41 @@ def add_zero_coverage(sample_name, sample_reference, nodupbam, hapall, zero_cove
     total_coverage = total_length - total_zero_coverage
     genome_coverage = "{:.2%}".format(total_coverage / total_length)
 
-    header_out = open('v_header.csv', 'w+')
-    with open(hapall) as fff:
-        for line in fff:
-            if re.search('^#', line):
-                print(line.strip(), file=header_out)
-    header_out.close()
-
     vcf_df = pd.read_csv(hapall, sep='\t', header=None, names=["CHROM", "POS", "ID", "REF", "ALT", "QUAL", "FILTER", "INFO", "FORMAT", "Sample"], comment='#')
-
     good_snp_count = len(vcf_df[(vcf_df['ALT'].str.len() == 1) & (vcf_df['REF'].str.len() == 1) & (vcf_df['QUAL'] > 150)])
 
-    vcf_df_snp = vcf_df[vcf_df['REF'].str.len() == 1]
-    vcf_df_snp = vcf_df_snp[vcf_df_snp['ALT'].str.len() == 1]
-    vcf_df_snp['ABS_VALUE'] = vcf_df_snp['CHROM'].map(str) + '-' + vcf_df_snp['POS'].map(str)
-    vcf_df_snp = vcf_df_snp.set_index('ABS_VALUE')
-    cat_df = pd.concat([vcf_df_snp, zero_df], axis=1, sort=False)
-    cat_df = cat_df.drop(columns=['CHROM', 'POS', 'depth'])
-    cat_df[['ID', 'ALT', 'QUAL', 'FILTER', 'INFO']] = cat_df[['ID', 'ALT', 'QUAL', 'FILTER', 'INFO']].fillna('.')
-    cat_df['REF'] = cat_df['REF'].fillna('N')
-    cat_df['FORMAT'] = cat_df['FORMAT'].fillna('GT')
-    cat_df['Sample'] = cat_df['Sample'].fillna('./.')
-    cat_df['temp'] = cat_df.index.str.split('-')
-    cat_df[['CHROM', 'POS']] = pd.DataFrame(cat_df.temp.values.tolist(), index=cat_df.index)
-    cat_df = cat_df[['CHROM', 'POS', 'ID', 'REF', 'ALT', 'QUAL', 'FILTER', 'INFO', 'FORMAT', 'Sample']]
-    cat_df['POS'] = cat_df['POS'].astype(int)
-    cat_df = cat_df.sort_values(['CHROM', 'POS'])
-    cat_df.to_csv('v_annotated_body.csv', sep='\t', header=False, index=False)
-    cat_files = ['v_header.csv', 'v_annotated_body.csv']
-    with open(zero_coverage_vcf, "wb") as outfile:
-        for cf in cat_files:
-            with open(cf, "rb") as infile:
-                outfile.write(infile.read())
+    if total_zero_coverage > 0:
+        header_out = open('v_header.csv', 'w+')
+        with open(hapall) as fff:
+            for line in fff:
+                if re.search('^#', line):
+                    print(line.strip(), file=header_out)
+        header_out.close()
+        vcf_df_snp = vcf_df[vcf_df['REF'].str.len() == 1]
+        vcf_df_snp = vcf_df_snp[vcf_df_snp['ALT'].str.len() == 1]
+        vcf_df_snp['ABS_VALUE'] = vcf_df_snp['CHROM'].map(str) + '-' + vcf_df_snp['POS'].map(str)
+        vcf_df_snp = vcf_df_snp.set_index('ABS_VALUE')
+        cat_df = pd.concat([vcf_df_snp, zero_df], axis=1, sort=False)
+        cat_df = cat_df.drop(columns=['CHROM', 'POS', 'depth'])
+        cat_df[['ID', 'ALT', 'QUAL', 'FILTER', 'INFO']] = cat_df[['ID', 'ALT', 'QUAL', 'FILTER', 'INFO']].fillna('.')
+        cat_df['REF'] = cat_df['REF'].fillna('N')
+        cat_df['FORMAT'] = cat_df['FORMAT'].fillna('GT')
+        cat_df['Sample'] = cat_df['Sample'].fillna('./.')
+        cat_df['temp'] = cat_df.index.str.split('-')
+        cat_df[['CHROM', 'POS']] = pd.DataFrame(cat_df.temp.values.tolist(), index=cat_df.index)
+        cat_df = cat_df[['CHROM', 'POS', 'ID', 'REF', 'ALT', 'QUAL', 'FILTER', 'INFO', 'FORMAT', 'Sample']]
+        cat_df['POS'] = cat_df['POS'].astype(int)
+        cat_df = cat_df.sort_values(['CHROM', 'POS'])
+        cat_df.to_csv('v_annotated_body.csv', sep='\t', header=False, index=False)
+        cat_files = ['v_header.csv', 'v_annotated_body.csv']
+        with open(zero_coverage_vcf, "wb") as outfile:
+            for cf in cat_files:
+                with open(cf, "rb") as infile:
+                    outfile.write(infile.read())
+    else:
+        shutil.copyfile(hapall, zero_coverage_vcf)
     return (zero_coverage_vcf, good_snp_count, ave_coverage, genome_coverage)
-
+    
 
 def send_email_step1(email_list, runtime, path_found, summary_file):
     text = "See attached:  "
